stage4

+++ Wine_Datasete_02.py
@@ -9,7 +9,7 @@ class Wine(Dataset):
         # csv ファイルを読み込む。
         df = pd.read_csv(csv_path)
         data = np.array(df.iloc[:, 1:])  # データ (2 ~ 14列目)
-        labels = np.array(df.iloc[:, 0])  # ラベル (1列目)
+        labels = np.array(df.iloc[:, 0], dtype="long")  # ラベル (1列目)
         # データを標準化する。
         data = normalize(data)
         # クラス ID を 0 始まりにする。[1, 2, 3] -> [0, 1, 2]
